Add BlogImages and OldSlugs to AppDbContext
This commit introduces two new properties to the `AppDbContext` class: `BlogImages` and `OldSlugs`. Both properties are of type `DbSet`, allowing for the management of collections of `BlogImageEntity` and `BlogOldSlugEntity` objects. This enhancement improves the data model's functionality by facilitating the storage and retrieval of blog images and old slugs associated with blogs.